# src/main/java/com/example/Immobi/Dto/auth/AuthResponse.java
package com.example.Immobi.Dto.auth;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

/**
 * Author: QuanNH
 */
@Data
@NoArgsConstructor
@AllArgsConstructor
public class AuthResponse {
    private String token;
    private String username;
} 