# src/main/java/com/example/Immobi/Dto/auth/LoginRequest.java
package com.example.Immobi.Dto.auth;

import jakarta.validation.constraints.NotBlank;
import lombok.Data;

/**
 * Author: QuanNH
 */
@Data
public class LoginRequest {
    
    @NotBlank(message = "Username is required")
    private String username;
    
    @NotBlank(message = "Password is required")
    private String password;
} 