# swagger.json
{"openapi":"3.1.0","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:9002","description":"Generated server url"}],"tags":[{"name":"Authentication","description":"Authentication management API"},{"name":"Number Guessing Game","description":"API endpoints for number guessing game"},{"name":"Player","description":"API endpoints for player information and leaderboard"}],"paths":{"/api/players/refresh-profile":{"post":{"tags":["Player"],"summary":"Refresh profile cache","description":"Force refresh current user's profile cache","operationId":"refreshProfileCache","responses":{"200":{"description":"Cache refreshed successfully","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BaseResponseVoid"}}}},"401":{"description":"Not authenticated","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BaseResponseVoid"}}}}}}},"/api/players/admin/reset-leaderboard":{"post":{"tags":["Player"],"summary":"Reset leaderboard","description":"Force rebuild leaderboard from database (admin only)","operationId":"resetLeaderboard","responses":{"200":{"description":"Leaderboard reset successfully","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BaseResponseVoid"}}}},"403":{"description":"Access denied","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BaseResponseVoid"}}}}}}},"/api/game/reset":{"post":{"tags":["Number Guessing Game"],"summary":"Reset turns","description":"Reset player's turns to default value","operationId":"resetTurns","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BaseResponseVoid"}}}}}}},"/api/game/guess":{"post":{"tags":["Number Guessing Game"],"summary":"Make a guess","description":"Guess a number between 1 and 5","operationId":"makeGuess","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GuessRequest"}}},"required":true},"responses":{"200":{"description":"Guess processed successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponse"}}}},"400":{"description":"Invalid input or no turns left","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BaseResponseGuessResponse"}}}},"401":{"description":"Not authenticated","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BaseResponseGuessResponse"}}}}}}},"/api/game/buy-turns":{"post":{"tags":["Number Guessing Game"],"summary":"Buy additional turns","description":"Purchase 5 more turns","operationId":"buyTurns","responses":{"200":{"description":"Turns purchased successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponse"}}}},"401":{"description":"Not authenticated","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BaseResponseBuyTurnsResponse"}}}},"402":{"description":"Payment required/failed","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BaseResponseBuyTurnsResponse"}}}}}}},"/api/auth/register":{"post":{"tags":["Authentication"],"summary":"User Register","description":"Creates a new user account and returns authentication token","operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterRequest"}}},"required":true},"responses":{"201":{"description":"User registered successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponse"}}}},"400":{"description":"Invalid input or username already exists","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BaseResponseAuthResponse"}}}}}}},"/api/auth/login":{"post":{"tags":["Authentication"],"summary":"User login","description":"Validates credentials and returns authentication token","operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"Login successful","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponse"}}}},"401":{"description":"Invalid credentials","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BaseResponseAuthResponse"}}}}}}},"/api/players/me":{"get":{"tags":["Player"],"summary":"Get profile","description":"Get current player profile information","operationId":"getPlayerProfile","responses":{"200":{"description":"Profile retrieved successfully","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BaseResponsePlayerProfileDto"}}}},"401":{"description":"Not authenticated","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BaseResponsePlayerProfileDto"}}}}}}},"/api/players/leaderboard":{"get":{"tags":["Player"],"summary":"Get leaderboard","description":"Get top players sorted by score","operationId":"getLeaderboard","parameters":[{"name":"limit","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}}],"responses":{"200":{"description":"Leaderboard retrieved successfully","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BaseResponseListLeaderboardEntryDto"}}}}}}}},"components":{"schemas":{"BaseResponseVoid":{"type":"object","properties":{"success":{"type":"boolean"},"code":{"type":"integer","format":"int32"},"message":{"type":"string"},"data":{},"timestamp":{"type":"string","format":"date-time"}}},"GuessRequest":{"type":"object","properties":{"number":{"type":"integer","format":"int32","maximum":5,"minimum":1}},"required":["number"]},"BaseResponse":{"type":"object","properties":{"success":{"type":"boolean"},"code":{"type":"integer","format":"int32"},"message":{"type":"string"},"data":{},"timestamp":{"type":"string","format":"date-time"}}},"BaseResponseGuessResponse":{"type":"object","properties":{"success":{"type":"boolean"},"code":{"type":"integer","format":"int32"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/GuessResponse"},"timestamp":{"type":"string","format":"date-time"}}},"GuessResponse":{"type":"object","properties":{"correct":{"type":"boolean"},"generatedNumber":{"type":"integer","format":"int32"},"guessedNumber":{"type":"integer","format":"int32"},"remainingTurns":{"type":"integer","format":"int32"},"score":{"type":"integer","format":"int32"},"message":{"type":"string"}}},"BaseResponseBuyTurnsResponse":{"type":"object","properties":{"success":{"type":"boolean"},"code":{"type":"integer","format":"int32"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/BuyTurnsResponse"},"timestamp":{"type":"string","format":"date-time"}}},"BuyTurnsResponse":{"type":"object","properties":{"successful":{"type":"boolean"},"newTurnCount":{"type":"integer","format":"int32"},"turnsAdded":{"type":"integer","format":"int32"},"currentScore":{"type":"integer","format":"int32"},"paymentMessage":{"type":"string"}}},"RegisterRequest":{"type":"object","properties":{"username":{"type":"string","maxLength":20,"minLength":3},"password":{"type":"string","maxLength":20,"minLength":6}},"required":["password","username"]},"AuthResponse":{"type":"object","properties":{"token":{"type":"string"},"username":{"type":"string"}}},"BaseResponseAuthResponse":{"type":"object","properties":{"success":{"type":"boolean"},"code":{"type":"integer","format":"int32"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/AuthResponse"},"timestamp":{"type":"string","format":"date-time"}}},"LoginRequest":{"type":"object","properties":{"username":{"type":"string","minLength":1},"password":{"type":"string","minLength":1}},"required":["password","username"]},"BaseResponsePlayerProfileDto":{"type":"object","properties":{"success":{"type":"boolean"},"code":{"type":"integer","format":"int32"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/PlayerProfileDto"},"timestamp":{"type":"string","format":"date-time"}}},"PlayerProfileDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"email":{"type":"string"},"username":{"type":"string"},"score":{"type":"integer","format":"int32"},"remainingTurns":{"type":"integer","format":"int32"},"rank":{"type":"integer","format":"int32"}}},"BaseResponseListLeaderboardEntryDto":{"type":"object","properties":{"success":{"type":"boolean"},"code":{"type":"integer","format":"int32"},"message":{"type":"string"},"data":{"type":"array","items":{"$ref":"#/components/schemas/LeaderboardEntryDto"}},"timestamp":{"type":"string","format":"date-time"}}},"LeaderboardEntryDto":{"type":"object","properties":{"userId":{"type":"integer","format":"int64"},"username":{"type":"string"},"score":{"type":"integer","format":"int32"},"rank":{"type":"integer","format":"int32"}}}}}}